# Options algorithm performance enhancement