--- a/Common/Data/Auxiliary/ZipEntryName.cs
+++ b/Common/Data/Auxiliary/ZipEntryName.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
  *
@@ -29,6 +29,15 @@ namespace QuantConnect.Data.Auxiliary
         public ZipEntryName()
         {
             DataType = MarketDataType.Auxiliary;
+        }
+
+        /// <summary>
+        /// Indicates whether this contains data that should be stored in the security cache
+        /// </summary>
+        /// <returns>Whether this contains data that should be stored in the security cache</returns>
+        public override bool ShouldCache()
+        {
+            return false;
         }
 
         /// <summary>