<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->

Following @Martin-Molinero's find about the engine expending ~30% of the time of a simple options algorithm trying to store `ZipEntryName` data from the universe data into the security cache, this filters this data by allowing data types to indicate whether they should be cached into the security.

Performance for a simple options algorithm (run in release mode):

```csharp
public override void Initialize()
{
    SetStartDate(2021, 01, 01);
    SetEndDate(2022, 12, 31);
    
    AddEquity("GOOG");
    AddOption("GOOG");
}
```

Master:
```
20231102 19:18:59.024 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 100.73 seconds at 8531k data points per second. Processing total of 859,271,594 data points.
20231102 19:25:05.903 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 98.82 seconds at 8695k data points per second. Processing total of 859,271,594 data points.
20231102 19:28:06.534 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 97.38 seconds at 8824k data points per second. Processing total of 859,271,594 data points.
```

PR:
```
20231102 19:12:53.943 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 65.82 seconds at 13054k data points per second. Processing total of 859,271,594 data points.
20231102 19:15:09.710 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 65.73 seconds at 13073k data points per second. Processing total of 859,271,594 data points.
20231102 19:16:30.333 TRACE:: Debug: Algorithm Id:(BasicTemplateOptionsAlgorithm) completed in 62.39 seconds at 13772k data points per second. Processing total of 859,271,594 data points.
```

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

N/A

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->

N/A

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->

N/A

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

Manual runs of algorithms

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->